requirements.txt
tqdm==4.31.1
requests==2.21.0
asn1==2.2.0
urllib3==1.24.1
asn1crypto==0.24.0
pycrypto==2.6.1
